requirements.txt
django_celery
